// src/pages/thank-you/interfaces/paymentInfo.ts
export interface IPaymentInfos {
  vnpAmount: string | null;

  vnpBankCode: string | null;

  vnpBankTranNo: string | null;

  vnpCardType: string | null;

  vnpOrderInfo: string | null;

  vnpPayDate: string | null;

  vnpResponseCode: string | null;

  vnpTmnCode: string | null;

  vnpTransactionNo: string | null;

  vnpTransactionStatus: string | null;

  vnpTxnRef: string | null;

  vnpSecureHash: string | null;
}